[Joey-O] 1 .升级 Nunit 2.6到Nunit 3 2. 创建Enjoy.Core单元测试,.

+++ src/Orchard.Web/Modules/Enjoy.Core/Common/Extensions/AuditStatusExtension.cs
@@ -26,7 +26,7 @@ namespace Enjoy.Core
             }
             return "未定义的状态";
         }
-        public static string GenerateActionLink(this AuditStatus status, int id)
+        public static string GenerateActionLink(this AuditStatus status, long id)
         {
             switch (status)
             {
